requirements.txt
pydub
PyYAML